Move type definition away from polygraph.types
