+++ polygraph/types/nonnull.py
@@ -1,7 +1,5 @@
 from polygraph.exceptions import PolygraphValueError
-from polygraph.types.api import typedef
 from polygraph.types.basic_type import PolygraphType
-from polygraph.types.definitions import TypeKind
 from polygraph.types.type_builder import TypeBuilderMeta, type_builder_cache
 
 
@@ -11,7 +9,7 @@ class NonNull(PolygraphType, metaclass=TypeBuilderMeta):
     """
     @type_builder_cache
     def __new__(cls, type_):
-        type_name = typedef(type_).name
+        type_name = type_.__name__
 
         if issubclass(type, NonNull):
             raise TypeError("NonNull cannot modify NonNull types")
@@ -19,7 +17,6 @@ class NonNull(PolygraphType, metaclass=TypeBuilderMeta):
         class Type:
             name = type_name + "!"
             description = "A non-nullable version of {}".format(type_name)
-            kind = TypeKind.NON_NULL
             of_type = type_
 
         def __new_from_value__(cls, value):
