+++ polygraph/types/tests/test_enum.py
@@ -1,7 +1,6 @@
 from unittest import TestCase
 
 from polygraph.exceptions import PolygraphValueError
-from polygraph.types.api import typedef
 from polygraph.types.enum import EnumType, EnumValue
 
 
@@ -43,10 +42,6 @@ class EnumTest(TestCase):
 
         with self.assertRaises(PolygraphValueError):
             Shapes(Colours.BLUE)
-
-    def test_enum_type(self):
-        colour_type = typedef(Colours)
-        self.assertEqual(len(colour_type.enum_values), 3)
 
     def test_enum_value_name(self):
         class NamedValue(EnumType):
